requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==5.2
sqlparse==0.5.3
djangorestframework~=3.16.0
argparse~=1.4.0